fixed everything, app is rendering
