+++ client/src/components/Profile/ProfilePage.css
@@ -1,0 +1,4 @@
+body{
+    background-color: seagreen;
+    color: white;
+}
